1、修复细节问题 2、修复组合商品在单据中审核后金额错误。 3、单据明细记录tax_amount 4、采购单据支持组合商品 5、组合商品的数量只能是整数 6、所有单据的数量都支持小数 7、零售单据-数量支持小数 8、销售单据-数量支持小数 9、库存调整-数量支持小数 10、库存预警-数量支持小数 11、仓库调拨数量调整成小数 12、仓库调拨-数量支持小数 13、商品库存变动记录查询-数量支持小数 14、商品库存查询-数量支持小数 15、采购订单数量调整成小数 16、修改gitignore 17、订单中的数量从整数修改成小数 18、支持根据域名解析租户ID

+++ xingyun-sc/src/main/java/com/lframework/xingyun/sc/bo/purchase/QueryPurchaseOrderBo.java
@@ -73,13 +73,13 @@ public class QueryPurchaseOrderBo extends BaseBo<PurchaseOrder> {
    * 采购数量
    */
   @ApiModelProperty("采购数量")
-  private Integer totalNum;
+  private BigDecimal totalNum;
 
   /**
    * 赠品数量
    */
   @ApiModelProperty("赠品数量")
-  private Integer totalGiftNum;
+  private BigDecimal totalGiftNum;
 
   /**
    * 采购金额
