+++ xingyun-sc/src/main/java/com/lframework/xingyun/sc/bo/sale/returned/GetSaleReturnBo.java
@@ -103,13 +103,13 @@ public class GetSaleReturnBo extends BaseBo<SaleReturnFullDto> {
    * 退货数量
    */
   @ApiModelProperty("退货数量")
-  private Integer totalNum;
+  private BigDecimal totalNum;
 
   /**
    * 赠品数量
    */
   @ApiModelProperty("赠品数量")
-  private Integer giftNum;
+  private BigDecimal giftNum;
 
   /**
    * 退货金额
@@ -299,19 +299,19 @@ public class GetSaleReturnBo extends BaseBo<SaleReturnFullDto> {
      * 出库数量
      */
     @ApiModelProperty("出库数量")
-    private Integer outNum;
+    private BigDecimal outNum;
 
     /**
      * 剩余退货数量
      */
     @ApiModelProperty("剩余退货数量")
-    private Integer remainNum;
+    private BigDecimal remainNum;
 
     /**
      * 退货数量
      */
     @ApiModelProperty("退货数量")
-    private Integer returnNum;
+    private BigDecimal returnNum;
 
     /**
      * 原价
@@ -402,7 +402,7 @@ public class GetSaleReturnBo extends BaseBo<SaleReturnFullDto> {
             dto.getOutSheetDetailId());
         this.outNum = outSheetDetail.getOrderNum();
         this.remainNum = NumberUtil.sub(outSheetDetail.getOrderNum(),
-            outSheetDetail.getReturnNum()).intValue();
+            outSheetDetail.getReturnNum());
       }
     }
   }
