+++ xingyun-sc/src/main/java/com/lframework/xingyun/sc/entity/PreTakeStockSheetDetail.java
@@ -3,6 +3,7 @@ package com.lframework.xingyun.sc.entity;
 import com.baomidou.mybatisplus.annotation.TableName;
 import com.lframework.starter.web.core.dto.BaseDto;
 import com.lframework.starter.web.core.entity.BaseEntity;
+import java.math.BigDecimal;
 import lombok.Data;
 
 /**
@@ -36,17 +37,17 @@ public class PreTakeStockSheetDetail extends BaseEntity implements BaseDto {
   /**
    * 初盘数量
    */
-  private Integer firstNum;
+  private BigDecimal firstNum;
 
   /**
    * 复盘数量
    */
-  private Integer secondNum;
+  private BigDecimal secondNum;
 
   /**
    * 抽盘数量
    */
-  private Integer randNum;
+  private BigDecimal randNum;
 
   /**
    * 排序
