+++ xingyun-sc/src/main/java/com/lframework/xingyun/sc/entity/SaleOrder.java
@@ -52,12 +52,12 @@ public class SaleOrder extends BaseEntity implements BaseDto {
   /**
    * 销售数量
    */
-  private Integer totalNum;
+  private BigDecimal totalNum;
 
   /**
    * 赠品数量
    */
-  private Integer totalGiftNum;
+  private BigDecimal totalGiftNum;
 
   /**
    * 销售金额
