+++ xingyun-sc/src/main/java/com/lframework/xingyun/sc/excel/purchase/returned/PurchaseReturnExportModel.java
@@ -70,13 +70,13 @@ public class PurchaseReturnExportModel extends BaseBo<PurchaseReturn> implements
    * 商品数量
    */
   @ExcelProperty("商品数量")
-  private Integer receiveNum;
+  private BigDecimal receiveNum;
 
   /**
    * 赠品数量
    */
   @ExcelProperty("赠品数量")
-  private Integer giftNum;
+  private BigDecimal giftNum;
 
   /**
    * 操作时间
