+++ xingyun-sc/src/main/java/com/lframework/xingyun/sc/excel/sale/returned/SaleReturnExportModel.java
@@ -70,13 +70,13 @@ public class SaleReturnExportModel extends BaseBo<SaleReturn> implements ExcelMo
    * 商品数量
    */
   @ExcelProperty("商品数量")
-  private Integer totalNum;
+  private BigDecimal totalNum;
 
   /**
    * 赠品数量
    */
   @ExcelProperty("赠品数量")
-  private Integer giftNum;
+  private BigDecimal giftNum;
 
   /**
    * 操作时间
